+++ lib/src/extensions/null_to_empty_string.dart
@@ -1,0 +1,5 @@
+extension NullToEmptyString on Object? {
+  /// Identical to [toString] unless this is null, in which case this returns
+  /// the empty string.
+  String nullToEmptyString() => this?.toString() ?? '';
+}