+++ lib/src/zpl_elements/zpl_label.dart
@@ -1,22 +1,26 @@
-import 'zpl_element.dart';
+import 'package:equatable/equatable.dart';
 
-/// The "top level" [ZplElement] that is typically sent to a printer.
-class ZplLabel extends ZplElement {
+import 'basic.dart';
+
+/// Put many [ZplElement]s onto a single [ZplLabel].
+///
+/// This is what you typically send to the printer.
+class ZplLabel extends Equatable {
   const ZplLabel({
     required this.width,
     required this.height,
     required this.children,
   });
 
-  // TODO: What units?
+  // Width of the label in dots.
   final int width;
 
-  // TODO: What units?
+  // Height of the label in dots.
   final int height;
 
   final List<ZplElement> children;
 
-  @override
+  /// The actual ZPL sent to the printer.
   String toZpl() {
     return '^XA${children.map((child) => child.toZpl()).join()}^XZ';
   }