+++ CHANGELOG.md
@@ -5,3 +5,12 @@ Add ZPL text and label.
 ## 0.1.1
 
 Add ZPL network printer.
+
+## 0.1.2
+
+- Change ZPL printer to be sealed
+- Add ZPL element equality
+
+## 0.2.0
+
+- Add barcode element